model_server/pyproject.toml
[tool.poetry]
name = "metallocechevitca-model-server"
version = "0.1.0"
description = "susamo"
authors = ["Your Name <you@example.com>"]
license = "WTFPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ray = {extras = ["serve"], version = "^2.10.0"}
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
